feat: log the size of the different work queues every 15 minutes
When the Monitor has a lot of work to do, the queues fill up and it's hard to understand their size and progress.
For customers with larger clusters it can be very helpful to have some visibility into the queue sizes.

Logging an additional single line every 15 minutes, even if the queues are empty, should be more helpful than noisy.