feat: load user-defined Rego rules for workload auto-import
The Rego rules are loaded by snyk-monitor at the start of the application.
We want the rules to be provided client-side but any choice whether workloads should be imported or deleted happens in the upstream. This allows us to swap implementations, fix bugs, etc. in the back end without requiring any new versions to be deployed by users.

Test coverage provided in system tests to ensure the policy is being transmitted first thing on start up of the application.