fix: cache workloads using uid instead of name as caching key
This solves the case where a workload is added, deleted, and then added again. Snyk-monitor thinks the image for this workload has already been scanned (because we cache this internally by name), however, this is a completely new workload and needs to be re-scanned so the generated Project can be visible for import in Snyk. So use uid in the caching key instead of name, which ensures a completely new instance of the new workload will result in a new cache key.