fix: load Rego policy from a different location
The Rego policy was mounted in the same location as the image temporary storage. Snyk Monitor tries to clean up this directory on start up (for the case where we get storage mounted with a bunch of extra files in it), so it also tried to delete the policies location. This resulted in an error and is also a mistake.

Swap the location from /var/tmp to /tmp.