fix: upgrade K8s client library to add Keep Alive to watches
This has been an outstanding issue for over a month, especially prevalent in AKS. After some inactivity in the cluster, the persistent connection to the K8s API server got interrupted and was never able to recover.

This new fix in the K8s client library puts a 30s keep alive to the Watch connection so that it does not get interrupted.