chore: remove --generator flag for kubectl in tests
We deploy a Pod in our integration tests using the --generator flag of kubectl.
Since 1.18 of kubectl, it seems that this flag has been deprecated and is no longer used: "has no effect and will be removed in the future".

This was caught by our tests because we try to use the latest version of kubectl, so breaking changes like this will surface.