fix: revert checking for parent workloads being controllers
This reverts the change in 07f4f34c8fb7a79f8e04e098f91d66015c99a13b.
DeploymentConfigs on OpenShift create ReplicationControllers which in turn create Pods. However, the ReplicationControllers in this case are not marked as controllers, hence we cannot walk the list of owner references and can never scan DeploymentConfigs.

This change ensures we do not strictly check for controllers to allow us to read DeploymentConfigs.