chore: pin Operator Lifecycle Manager to version 0.17.0
The recent release of version 0.18.0 releases a breaking change and causes the setup of our OLM tests to fail.
For now instead of using the main branch, we pin to the non-breaking change release until we understand what we need to change in our test setup to accommodate the new version.