fix: use Recreate strategy for snyk-monitor when using PVC as storage
When using a PersistentVolumeClaim, it has a Volume attached to it that can be shared by Pods only on the same worker node.

On upgrades to the Deployment, Kubernetes assigns the new Pods to random Kubernetes worker nodes. If the new Pod is assigned to a different worker node, the PVC cannot be detached from the old Pod and assigned to the new Pod.

This is why when we use a PVC we now change the snyk-monitor Deployment strategy to Recreate. Kubernetes will first delete the old Pod, freeing up the PVC, which can then be used in the new Pod.