fix: don't attempt to log an entire pod object
When encountering a pod associated with an owner, but not finding its owner, it should suffice to log the pod's metadata so we know the name and ownerReference.
Logging the entire pod object could result in either accidentally leaking sensitive data, or just logging an unbounded huge object and break log parsing.