fix: the snyk-operator does not use OCP 4.9 due to deprecated APIs
Kubernetes has been deprecating API(s) which will be removed and no longer available in 1.22 and in the Openshift version 4.9.

For now we disable upgrades until we resolve the deprecated APIs and upgrade them to supported versions. Without this change the current version of the Operator does not get approved to the Community Operators repo.
https://github.com/redhat-openshift-ecosystem/community-operators-prod/pull/208#issuecomment-915354409