Add init delay to liveness and readiness probes
We are attempting to use the monitor and in our environment are experiencing probes failing on startup. From our investigation we see the monitor pinning the CPU on startup which is causing the probes to fail and in turn the container to be killed before it starts up.

I was able to get things to startup sometimes by giving the container more CPU cores but even this was not reliable. Adding an init delay should hopefully resolve this and allow the container to start before it is nuked due to failing probes.

As the probes are not overly useful here I just went with a long init delay.