feat: allow configuring skopeo compression and image scan worker count
Allowing to experiment with these values to see if it helps alleviate performance issues.
We see that in some cases the CPU is pegged to 100% and causes lots of throttling by Kubernetes.

We suspect that part of the problem is the heavy compression done by skopeo which is used to pull an image from the container registry and then store it to disk.
Additionally, allow to configure the number of workers that run in parallel. This would allow to further reduce the CPU load.