fix: remove backup option to pull oci archive
Skopeo is capable of pulling an OCI image and it doesn't need a backup to pull an OCI archive. If the Docker archive pulling fails for some reason, the OCI archive pulling won't be able to succeed either. This backup option was introduced by mistake thinking that skopeo cannot handle certain images as Docker archives. The underlying problem was that the image itself was not OCI compliant and couldn't be pulled, so OCI archive pulling wouldn't have helped.