feat: remove workloads from scan queue when they are deleted in K8s
Up until now the queue of workloads to scan would keep growing indefinitely as we only appended new workloads to it.
However, once workloads are deleted from Kubernetes, they no longer need to be in the queue.

This didn't cause any issues with trying to process deleted workloads, because the queue handler would detect that the workload is gone and continue processing other items. However, it created alerts that the queue grows indefinitely and caused the snyk-monitor to process some work that was never going to succeed.