fix: reduce workers count to 5
On some customer deployments there is a really high spike of CPU usage that causes healthcheck failures. Reducing the worker count should reduce the CPU usage and make this less of a problem.
This change is not expected to affect scanning speed and make it slower as most of the time is spent on pulling images.