feat: switch to alpine base image to reduce vulns
We have made the decision to publish the snyk-monitor with an Alpine base image, which has less packages and therefore less vulnerabilities.
We have recently been struck with a long case of high severity vulnerabilities in the UBI8 base image which had no fixes.

Some of our customers' pipelines break because we're publishing an image with high severity vulnerabilities. They therefore have to triage the vulns and also raise it with us to resolve.

This change affects the default Helm chart installation but overall should not have any performance impact or any problems with the product.

The Operator installation will NOT use Alpine - it will still use UBI8 because it is a requirement for running on Red Hat OpenShift and being a Certified product.