fix: managing cache state is now synchronous
We no longer use async when managing cache state because we suspect it spins the event loop and causes race conditions where the same image is scanned multiple times even though it should have been cached the first time.

Also, add debug logging to inspect the data in the cache.