fix: enqueue Kubernetes client API requests to prevent high load
When the snyk-monitor first starts, it receives Pod events for every Pod in the cluster. For each Pod, it tries to determine which workload runs it and loads the workload metadata via the Kubernetes API. There is no flow control here, so a burst of client requests are sent to the Kubernetes API. This occasionally results in rate limiting and thousands of concurrently open connections.

Adding a queue to send these requests only a couple at a time, reducing the load on the API server.