fix: cache workload details to reduce chances of rate limiting
On first start up, the snyk-monitor may do lots of Kubernetes API requests in order to load the workload metadata for all running workloads. We don't have a sophisticated logic for reading this data and in some cases we do the same request multiple times. For example, if a workload has 10 Pods, we will receive 10 events for these Pods and try to read the workload metadata 10 times.

We are now caching the workload metadata for a short period of time, easing the load on the API server and guarding against HTTP 429 from the server.