fix: use fastq as queueing library for API calls
The async.queue functionality does not preserve the error on promise rejections. Testing the fastq.promise functionality, it seems to do a better job and forwards the rejection correctly.

For now, replacing the queue only in the API clients (K8s API requests and Kubernetes Upstream requests).

The pod watch handler continues to use async because it ensures that promises are not thrown in the queue worker.